Representando valores monetários em centavos (Int); exibindo total gasto no ListaCard
- Migra preco e valorEstimado de Double para Int (centavos) em toda a stack
- Adiciona MIGRATION_2_3 recriando as tabelas com conversão ROUND(valor * 100)
- Extrai extrairQuantidadeNumerica e calcularTotalGasto para PrecoUtils.kt
- Exibe valor gasto real no ListaCard calculado a partir dos itens comprados

+++ app/src/main/java/com/example/nossafeira/data/model/ItemFeira.kt
@@ -23,7 +23,7 @@ data class ItemFeira(
     val nome: String,
     val quantidade: String,
     val categoria: Categoria,
-    val preco: Double = 0.0,
+    val preco: Int = 0,
     val comprado: Boolean = false,
     val criadoEm: Long = System.currentTimeMillis()
 )
